[Chore] #99 - KeywordData, KeywordDomain -> Base로 이관 및 기존 모듈 삭제
