+++ Projects/Data/BaseData/Sources/Keyword/DTO/Request/SearchKeywordRequest.swift
@@ -1,0 +1,14 @@
+//
+//  SearchKeywordRequest.swift
+//  BaseData
+//
+//  Created by Seoyeon Choi on 4/9/26.
+//  Copyright © 2026 kr.websoso.app. All rights reserved.
+//
+
+import Foundation
+import Networking
+
+struct SearchKeywordRequest: QueryItemConvertible {
+    let query: String
+}
