--- Projects/Data/KeywordData/Demo/KeywordDataDemoApp.swift
@@ -1,18 +1,0 @@
-//
-//  KeywordDataDemoApp.swift
-//  KeywordData
-//
-//  Created by Seoyeon Choi on 4/25/26.
-//
-
-import SwiftUI
-import KeywordData
-
-@main
-struct KeywordDataDemoApp: App {
-    var body: some Scene {
-        WindowGroup {
-            KeywordDataDemoView()
-        }
-    }
-}
