--- Projects/Data/KeywordData/Project.swift
@@ -1,24 +1,0 @@
-//
-//  Project.swift
-//  AppManifests
-//
-//  Created by Seoyeon Choi on 4/9/26.
-//
-
-import Foundation
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-import DependencyPlugin
-
-let project = Project.createDataModule(
-    name: ModuleType.data(.keyword).name,
-    targets: [.sources, .demo, .testing, .tests],
-    internalDependencies: [
-        .module(.core(.networking)),
-        .module(.core(.logger)),
-        .module(.data(.base)),
-        .module(.domain(.base)),
-        .module(.domain(.keyword))
-    ]
-)
