--- Projects/Data/KeywordData/Sources/DTO/Request/SearchKeywordRequest.swift
@@ -1,14 +1,0 @@
-//
-//  SearchKeywordRequest.swift
-//  KeywordData
-//
-//  Created by Seoyeon Choi on 4/9/26.
-//  Copyright © 2026 kr.websoso.app. All rights reserved.
-//
-
-import Foundation
-import Networking
-
-struct SearchKeywordRequest: QueryItemConvertible {
-    let query: String
-}
