--- Projects/Data/KeywordData/Sources/Service/KeywordService.swift
@@ -1,13 +1,0 @@
-//
-//  KeywordService.swift
-//  KeywordData
-//
-//  Created by Seoyeon Choi on 4/9/26.
-//  Copyright © 2026 kr.websoso.app. All rights reserved.
-//
-
-import Foundation
-
-protocol KeywordService {
-    func searchKeyword(_ request: SearchKeywordRequest) async throws -> KeywordGroupsResponse
-}
