--- Projects/Data/KeywordData/Testing/KeywordTesting.swift
@@ -1,8 +1,0 @@
-//
-//  KeywordTesting.swift
-//  AppManifests
-//
-//  Created by Seoyeon Choi on 4/25/26.
-//
-
-import Foundation
