--- Projects/Data/KeywordData/Tests/KeywordDataTests.swift
@@ -1,8 +1,0 @@
-//
-//  KeywordDataTests.swift
-//  AppManifests
-//
-//  Created by Seoyeon Choi on 4/25/26.
-//
-
-import Foundation
