--- Projects/Domain/KeywordDomain/Project.swift
@@ -1,16 +1,0 @@
-//
-//  Project.swift
-//  Manifests
-//
-//  Created by Seoyeon Choi on 2/6/26.
-//
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-import DependencyPlugin
-
-let project = Project.createDomainModule(
-    name: ModuleType.domain(.keyword).name,
-    targets: [.sources, .testing, .tests],
-    internalDependencies: [.module(.domain(.base))]
-)
